move simulate_header to layout.rs
